% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{synthetic_expression}
\alias{synthetic_expression}
\title{Synthetic Expression of Two Distinct Classes}
\format{
\subsection{\code{synthetic_expression}}{

A list containing two items:
\describe{
\item{expression}{Normalized log expression of 12 samples across 100 genes}
\item{classes}{A factor vector of classes of each of the 12 samples}
}
}
}
\usage{
synthetic_expression
}
\description{
Synthetic Expression of Two Distinct Classes
}
\keyword{datasets}
