% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clanc-predict.R
\name{predict.clanc}
\alias{predict.clanc}
\title{Predict from a \code{clanc}}
\usage{
\method{predict}{clanc}(object, new_data, type, assay = NULL, ...)
}
\arguments{
\item{object}{A \code{clanc} object.}

\item{new_data}{A data frame or matrix of new predictors.}

\item{type}{A single character. The type of predictions to generate.
Valid options are:
\itemize{
\item \code{"numeric"} for numeric predictions.
}}

\item{assay}{If \code{object} inherits \code{SummarizedExperiment}, the index of the
assay.}

\item{...}{Not used, but required for extensibility.}

\item{method}{If \code{type} is \code{numeric}, the method of correlation}
}
\value{
A tibble of predictions. The number of rows in the tibble is guaranteed
to be the same as the number of rows in \code{new_data}.
}
\description{
Predict from a \code{clanc}
}
